1.26.45 support (#3)
* 26.40 support

* update level-chunk-packet

* feat: 26.40 support

* feat: 1.26.44 support

* bump protocol version

* feat: update game data

* feat: 1.26.45 support

+++ src/enums/permission-level.zig
@@ -1,4 +1,4 @@
-pub const PermissionLevel = enum(i32) {
+pub const PermissionLevel = enum(i8) {
     Visitor = 0,
     Member = 1,
     Operator = 2,
